+++ src/tests/compiles.cpp
@@ -1,8 +1,0 @@
-#include "../headers/Mappings.hpp"
-#include "../headers/DMs.hpp"
-#include "../headers/Envelope.hpp"
-#include "../headers/User.hpp"
-
-int main() {
-    return 0;
-}