+++ tests/compiles.cpp
@@ -1,0 +1,9 @@
+#include "../src/headers/Mappings.hpp"
+#include "../src/headers/DMs.hpp"
+#include "../src/headers/Envelope.hpp"
+#include "../src/headers/User.hpp"
+#include "../src/headers/Time.hpp"
+
+int main() {
+    return 0;
+}